+++ isy-security/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `IFS-4591`: Hinzufügen von Authentifizierungsmethoden zur Authentifizierung von Clients und Systemen ohne Issuer-URI.
 - `IFS-4754`: Einführung von Caching im Authentifizierungsprozess
+- `IFS-4785`: Hinzufügen einer Property für die Restlebensdauer gecachter OAuth2-Token
 - `IFS-4752`: Wiederherstellen der initialen Authentication nach Authentifizierung mit @Authenticate-Annotation
 
 # 4.0.0